pyproject.toml
[tool.poetry]
name = "lab-fortune"
version = "0.1.0"
description = ""
authors = ["emma <eqbrown@isf.edu.hk>"]
readme = "README.md"
packages = [{include = "lab_fortune"}]

[tool.poetry.dependencies]
python = "^3.10"
django-banjo = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
